angstromctf/2020/crypto/discrete-superlog/requirements.txt
git+git://github.com/elliptic-shiho/primefac-fork@master
pycrypto==2.6.1